asm-verify: fix signature mismatches incl. Touch float coords (#266)
From the signature-mismatch round (verified vs v1.6.1): the port carried several
function signatures that diverged from the binary, so asm-verify never paired
them. Fix the real ones:

- Mortar::Touch: coords were truncated to int at the storage site
  (___UpdateInternal cast SDL float coords to int32_t), losing sub-pixel blade
  precision. TouchState x/y -> float; GetTouchPos/GetTouchDelta -> (uint, float&,
  float&); GetTouchInReigion -> (float,float,float,float); InputDeviceBada locals
  -> float. (Touch::GetTouchPos @0x002429d4 / UpdateInternal @0x00242868.)
- MortarSound::SetPitch(unsigned int) -> (float) (@0x00230218; VFP ABI).
- Math::Random::Seed(uint64) -> (uint32) + reset m_Mult/m_Inc (@0x0012c998;
  verified no determinism change -- the only seed source is 32-bit).
- MatrixManager::SetupOrtho/SetupLookAt: re-add the defaulted out-matrix param
  the binary takes (@0x002571d0/@0x002572ac), matching SetupPerspective.
- SplatEntity::Init(): add the (void*,long,Vec3*) vtable args the binary declares
  (ignored, @0x001eb264).

Tests: test_utf8_text stub updated for the new MatrixManager signatures;
test_random seeds with a 32-bit value (ctor stays 64-bit / Seed is 32-bit, as in
the binary).

+++ src/engine/input/InputDeviceBada.cpp
@@ -60,8 +60,8 @@ void InputDeviceBada::Update(float /*dt*/) {
 
     ++m_EventStamp;
 
-    int posX = 0;
-    int posY = 0;
+    float posX = 0.0f;
+    float posY = 0.0f;
 
     if (m_ActiveTouchId == 0) {
         // No active touch yet: try to acquire one this frame.
@@ -77,10 +77,10 @@ void InputDeviceBada::Update(float /*dt*/) {
         }
         // Freshly acquired touch -> emit current position (press-edge).
         touch.GetTouchPos(m_ActiveTouchId, posX, posY);
-        AxisEvent(0x74, 0x20, (float)posX,
-                  (float)(posX - (int)m_LastTouchX), m_EventStamp, 0);
-        AxisEvent(0x75, 0x20, (float)posY,
-                  (float)(posY - (int)m_LastTouchY), m_EventStamp, 0);
+        AxisEvent(0x74, 0x20, posX,
+                  posX - (float)m_LastTouchX, m_EventStamp, 0);
+        AxisEvent(0x75, 0x20, posY,
+                  posY - (float)m_LastTouchY, m_EventStamp, 0);
         // mask 1 = press-edge.
         ButtonPressed(0x6c, 1, 1.0f, m_EventStamp, 0);
         m_LastTouchX = (uint32_t)posX;
@@ -93,8 +93,8 @@ void InputDeviceBada::Update(float /*dt*/) {
             recent = touch.GetAnyTouch();
         }
         uint32_t tracked = m_ActiveTouchId;
-        int dX = 0;
-        int dY = 0;
+        float dX = 0.0f;
+        float dY = 0.0f;
         bool active = touch.GetTouchDelta(m_ActiveTouchId, dX, dY) != 0;
         if (!active || recent != tracked) {
             // Lost/changed touch: emit release (mask 4) then up (mask 8),
@@ -105,10 +105,10 @@ void InputDeviceBada::Update(float /*dt*/) {
             goto callbacks;
         }
         touch.GetTouchPos(m_ActiveTouchId, posX, posY);
-        AxisEvent(0x74, 0x20, (float)posX,
-                  (float)(posX - (int)m_LastTouchX), m_EventStamp, 0);
-        AxisEvent(0x75, 0x20, (float)posY,
-                  (float)(posY - (int)m_LastTouchY), m_EventStamp, 0);
+        AxisEvent(0x74, 0x20, posX,
+                  posX - (float)m_LastTouchX, m_EventStamp, 0);
+        AxisEvent(0x75, 0x20, posY,
+                  posY - (float)m_LastTouchY, m_EventStamp, 0);
         // mask 2 = held/move.
         ButtonPressed(0x6c, 2, 1.0f, m_EventStamp, 0);
         m_LastTouchX = (uint32_t)posX;
